Integrated Home component to Games component
