+++ client/src/components/Games/Games.css
@@ -1,3 +1,18 @@
+.games-wrapper {
+  display: -webkit-box;
+  display: -ms-flexbox;
+  display: flex;
+  -webkit-box-orient: vertical;
+  -webkit-box-direction: normal;
+      -ms-flex-direction: column;
+          flex-direction: column;
+  -ms-flex-line-pack: center;
+      align-content: center;
+  min-height: 90vh;
+  background-color: #e9a197;
+  margin: 0;
+}
+
 .games {
   width: 80%;
   margin: 20px 0 20px 0;
