--- client/src/components/home/Home.css
@@ -1,15 +1,0 @@
-.home {
-  display: -webkit-box;
-  display: -ms-flexbox;
-  display: flex;
-  -webkit-box-orient: vertical;
-  -webkit-box-direction: normal;
-      -ms-flex-direction: column;
-          flex-direction: column;
-  -ms-flex-line-pack: center;
-      align-content: center;
-  min-height: 90vh;
-  background-color: #e9a197;
-  margin: 0;
-}
-/*# sourceMappingURL=Home.css.map */
