--- client/src/components/home/Home.scss
@@ -1,10 +1,0 @@
-@import '../../styles/Variables.scss';
-
-.home {
-  display: flex;
-  flex-direction: column;
-  align-content: center;
-  min-height: 90vh;
-  background-color: $secondary-color;
-  margin: 0;
-}
